requirements.txt
Django==2.1.5
beautifulsoup4==4.8.0
cachetools==3.1.1
certifi==2019.9.11
chardet==3.0.4
cycler==0.10.0
djangorestframework==3.10.3
google==2.0.2
google-api-core==1.14.2
google-auth==1.6.3
google-cloud-vision==0.39.0
googleapis-common-protos==1.6.0
grpcio==1.23.0
idna==2.8
imutils==0.5.3
kiwisolver==1.1.0
matplotlib==3.1.1
numpy==1.17.2
opencv-python==4.1.1.26
pandas==0.25.1
pip==19.0.3
protobuf==3.9.1
pyasn1==0.4.7
pyasn1-modules==0.2.6
pyparsing==2.4.2
python-dateutil==2.8.0
pytz==2019.2
requests==2.22.0
rsa==4
scipy==1.3.1
setuptools==40.8.0
six==1.12.0
soupsieve==1.9.3
sqlparse==0.3.0
urllib3==1.25.3
mysqlclient==1.4.4 
